requirements.txt
asgiref==3.7.2
Django==3.1.5
django-restframework==0.0.1
djangorestframework==3.14.0
gunicorn==20.0.4
psycopg2-binary==2.9.6
pycparser==2.20
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
